requirements.txt
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.40.0
humanfriendly==10.0
kiwisolver==1.4.4
matplotlib==3.7.2
mpmath==1.3.0
numpy==1.25.1
onnx==1.14.0
onnxruntime==1.15.1
packaging==23.1
Pillow==10.0.0
protobuf==4.23.4
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
six==1.16.0
sympy==1.12
typing_extensions==4.7.1
